# Exit light calculation early when pixel outside of light bounding rectangle