Potential fix for https://github.com/godotengine/godot/issues/81152
Fixes: https://github.com/godotengine/godot/issues/90905

Right now when rendering a CanvasItem for each pixel in the fragment shader we loop over all lights that intersect that CanvasItem and calculate their contribution to the pixel. If the pixel is outside of the light's texture, we set the alpha value of the final result to 0 which results in that light not contributing to the pixel. 

Unfortunately, this means that we read from the light's texture still and run the light shader for that light, even though it won't have any effect. 

Normally, using a continue op in a loop is not a good idea is it can stop the compiler from doing certain beneficial optimizations. But in this case, the compiler can't optimize this loop for many other reasons, so this ends up being a net win. 

**Performance testing**
|        | iGPU RD | iGPU compatibility | Pixel 4 RD      | Pixel 4 compatibility |
|--------|---------|--------------------|-----------------|-----------------------|
| Before | 240 FPS | 140 FPS            | 42 FPS          | 50 FPS                |
| After  | 370 FPS | 275 FPS            | 60 FPS (locked) | 60 FPS (locked)       |
* Note the Pixel 4 does not support not using VSync, so it is unclear what the actual performance gain is as the device caps at 60 FPS.

**Further testing**

This PR would benefit from being tested on even lower end mobile, on web export, ~and on a desktop device with a dedicated GPU~ (done)

[Light2DTestUpdated.zip](https://github.com/godotengine/godot/files/15046194/Light2DTestUpdated.zip)