@tdaven 2024-04-23
>     1. I couldn't figure out how to benchmark this one reliably (I can use Xephyr but this breaks any performance expectations due to its own overhead). `LIBGL_ALWAYS_SOFTWARE=1` does not work on NVIDIA; it only has an effect on Mesa drivers. [↩](#user-content-fnref-1-33b3f095a6e19db5ec7ccee2585c9c85)

If your system is using glvnd and EGL you might be able to use this:
https://gitlab.freedesktop.org/glvnd/libglvnd/-/blob/master/src/EGL/icd_enumeration.md?ref_type=heads